requirements.txt
dash==2.10.1
dash-bootstrap-components==1.4.1
pandas>=2.2.0
requests==2.31.0
plotly==5.16.0
python-dateutil==2.8.2
python-dotenv
